requirements.txt

ollama==0.4.4
openai==1.56.2
python-dotenv==1.0.1
tweepy==4.14.0
